pyproject.toml
[tool.poetry]
name = "property_manager"
version = "0.1.0"
description = "Tool for managing vacation rental properties."
authors = ["Thomas Stivers <thomas.stivers@gmail.com>"]
license = "GPL"

[tool.poetry.dependencies]
python = "^3.9"
Flask-AppBuilder = "^3.3.4"
usps-api = "^0.5"
Flask-Migrate = "^3.1.0"
mortgage = "^1.0.5"
mysqlclient = "^2.0.3"
Flask-Mail = "^0.9.1"
gunicorn = "^20.1.0"
unicorn = "^1.0.3"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
